Téléchargements/project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-Limiter==3.3.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyMySQL==1.1.0
cryptography==41.0.3